Move reducers for popups to src/reducers
